test(dropdown): spilt snapshot tests only for jest
